+++ unit_tests/migs/FV2204/segment_group_hierarchies/edi_energy_example_diagram.jpg
