+++ unit_tests/migs/FV2204/segment_group_hierarchies/sgh_iftsta.json
