+++ unit_tests/test_ahb_csv_reader.py
@@ -129,7 +129,7 @@ class TestAhbCsvReader:
         assert actual_code == expected_code
         assert actual_beschreibung == expected_beschreibung
 
-    @pytest.mark.datafiles("./unittests/ahbs/FV2204/UTILMD/11042.csv")
+    @pytest.mark.datafiles("./unit_tests/ahbs/FV2204/UTILMD/11042.csv")
     def test_csv_file_reading_11042(self, datafiles):
         path_to_csv: Path = datafiles / "11042.csv"
         reader = FlatAhbCsvReader(file_path=path_to_csv)
