+++ unit_tests/test_edifact_components.py
@@ -14,7 +14,7 @@ from maus.models.edifact_components import (
     SegmentSchema,
     ValuePoolEntry,
 )
-from unittests.serialization_test_helper import assert_serialization_roundtrip  # type:ignore[import]
+from unit_tests.serialization_test_helper import assert_serialization_roundtrip  # type:ignore[import]
 
 
 class TestEdifactComponents:
