+++ unit_tests/test_integration_21035.py
@@ -9,7 +9,7 @@ from maus.models.anwendungshandbuch import DeepAnwendungshandbuchSchema
 from maus.models.message_implementation_guide import SegmentGroupHierarchySchema
 from maus.reader.flat_ahb_reader import FlatAhbCsvReader
 from maus.reader.mig_reader import MigXmlReader
-from unittests.test_mig import ALL_SGH_FILES  # type:ignore[import]
+from unit_tests.test_mig import ALL_SGH_FILES  # type:ignore[import]
 
 
 class TestIntegration35001:
@@ -20,9 +20,9 @@ class TestIntegration35001:
 
     @ALL_SGH_FILES
     @ALL_MIG_XML_FILES
-    @pytest.mark.datafiles("./unittests/ahbs/FV2204/IFTSTA/21035.csv")
-    @pytest.mark.datafiles("./unittests/ahbs/FV2204/IFTSTA/21035_deep.json")
-    @pytest.mark.datafiles("./unittests/ahbs/FV2204/IFTSTA/21035_maus.json")
+    @pytest.mark.datafiles("./unit_tests/ahbs/FV2204/IFTSTA/21035.csv")
+    @pytest.mark.datafiles("./unit_tests/ahbs/FV2204/IFTSTA/21035_deep.json")
+    @pytest.mark.datafiles("./unit_tests/ahbs/FV2204/IFTSTA/21035_maus.json")
     def test_csv_file_reading_21035(self, datafiles):
         path_to_csv: Path = datafiles / "21035.csv"
         reader = FlatAhbCsvReader(file_path=path_to_csv)
