+++ unit_tests/test_maus.py
@@ -13,7 +13,7 @@ from maus.models.edifact_components import (
     ValuePoolEntry,
 )
 from maus.models.message_implementation_guide import SegmentGroupHierarchy
-from unittests.serialization_test_helper import assert_serialization_roundtrip  # type:ignore[import]
+from unit_tests.serialization_test_helper import assert_serialization_roundtrip  # type:ignore[import]
 
 
 class TestMaus:
