+++ unit_tests/test_mig.py
@@ -3,13 +3,13 @@ from typing import List, Optional, Tuple
 import pytest  # type:ignore[import]
 
 from maus.models.message_implementation_guide import SegmentGroupHierarchy, SegmentGroupHierarchySchema
-from unittests.serialization_test_helper import assert_serialization_roundtrip  # type:ignore[import]
+from unit_tests.serialization_test_helper import assert_serialization_roundtrip  # type:ignore[import]
 
 ALL_SGH_FILES = pytest.mark.datafiles(
-    "./unittests/migs/FV2204/segment_group_hierarchies/sgh_mscons.json",
-    "./unittests/migs/FV2204/segment_group_hierarchies/sgh_utilmd.json",
-    "./unittests/migs/FV2204/segment_group_hierarchies/sgh_reqote.json",
-    "./unittests/migs/FV2204/segment_group_hierarchies/sgh_iftsta.json",
+    "./unit_tests/migs/FV2204/segment_group_hierarchies/sgh_mscons.json",
+    "./unit_tests/migs/FV2204/segment_group_hierarchies/sgh_utilmd.json",
+    "./unit_tests/migs/FV2204/segment_group_hierarchies/sgh_reqote.json",
+    "./unit_tests/migs/FV2204/segment_group_hierarchies/sgh_iftsta.json",
 )
