+++ unit_tests/test_mig_xml_reader_real_data.py
@@ -8,16 +8,16 @@ from maus.models.edifact_components import EdifactStackQuery
 from maus.reader.mig_reader import MigXmlReader
 
 ALL_MIG_XML_FILES = pytest.mark.datafiles(
-    "./unittests/migs/FV2204/template_xmls/mscons_1154.xml",
-    "./unittests/migs/FV2204/template_xmls/utilmd_1154.xml",
-    "./unittests/migs/FV2204/template_xmls/utilmd_2379.xml",
-    "./unittests/migs/FV2204/template_xmls/utilmd_2380.xml",
-    "./unittests/migs/FV2204/template_xmls/utilmd_7402.xml",
-    "./unittests/migs/FV2204/template_xmls/utilmd_3225.xml",
-    "./unittests/migs/FV2204/template_xmls/utilmd_9013.xml",
-    "./unittests/migs/FV2204/template_xmls/reqote.xml",
-    # "./unittests/migs/FV2204/template_xmls/utilmd.xml",
-    # "./unittests/migs/FV2204/template_xmls/iftsta.xml",
+    "./unit_tests/migs/FV2204/template_xmls/mscons_1154.xml",
+    "./unit_tests/migs/FV2204/template_xmls/utilmd_1154.xml",
+    "./unit_tests/migs/FV2204/template_xmls/utilmd_2379.xml",
+    "./unit_tests/migs/FV2204/template_xmls/utilmd_2380.xml",
+    "./unit_tests/migs/FV2204/template_xmls/utilmd_7402.xml",
+    "./unit_tests/migs/FV2204/template_xmls/utilmd_3225.xml",
+    "./unit_tests/migs/FV2204/template_xmls/utilmd_9013.xml",
+    "./unit_tests/migs/FV2204/template_xmls/reqote.xml",
+    # "./unit_tests/migs/FV2204/template_xmls/utilmd.xml",
+    # "./unit_tests/migs/FV2204/template_xmls/iftsta.xml",
 )
